backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-dotenv==1.0.0
django-filter==23.5
pandas==2.1.4
numpy==1.26.2


